Calculate particle offsets more reliably.
The previous implementation assumed `generateAnimationSubGeometries()` would be called multiple times. However, it turns out that in some cases (perhaps when the mesh is too far away to be rendered), it's only called once. In that case, the calculation needs to happen at the end of the one call.

But also, if it IS called multiple times, the second half of the function needs access to the up-to-date offsets, so I also recount them about halfway through.